requirements.txt
python-telegram-bot==13.7
# Python 3.8+
